\name{absorber-package}
\alias{absorber-package}
\docType{package}
\title{
\packageTitle{absorber}
}
\description{
absorber consists of two functions: "absorber.R" and "plot_selection.R".
For further information on how to use these functions, 
we refer the reader to the vignette of the package.
}
\details{
Two datasets are also provided within this package and used as examples of this manual and in the vignette. 
}
\author{
Mary E. Savino

Maintainer: Mary E. Savino <mary.savino@outlook.fr>
}
\references{
Savino, M. E. and Lévy-Leduc, C. (2024) A novel variable selection method in nonlinear multivariate models using B-splines with an application to geoscience. <https://hal.science/hal-04434820>.
}
\keyword{package}

