\name{x_obs}
\alias{x_obs}
\docType{data}
\title{
Observation matrix x of five variables
}
\description{
An example of 700 observations for the variable selection of function \eqn{f_1} (see Savino and Lévy-Leduc (2024) for more details) with five input variables.
}
\usage{data("x_obs")}
\format{
Numeric matrix of 700 rows and 5 columns.
}

