\name{y_obs}
\alias{y_obs}
\docType{data}
\title{
Values of the response variable of the noisy observation set of five input variables
}
\description{
An example of noisy observations obtained by adding a Gaussian noise to \eqn{f_1(x_i)} associated to the input values contained in x_obs.rda. See Savino and Lévy-Leduc (2024) for the expression of \eqn{f_1}.}
\usage{data("y_obs")}
\format{
  Numeric vector of 700 values.
}
